// projects/bible/web/src/interfaces/ISongResponse.ts
// Generated by https://quicktype.io

import { ISong } from './ISongsBook';

export interface ISongResponse {
  status: string;
  ok: boolean;
  data: ISongResponseData;
}

export interface ISongResponseData {
  data: FullBook;
}

export interface FullBook {
  _id: string;
  book: Book;
  views: number;
  likes: number;
  num: number;
  title: string;
  lyrics: string;
  videos: any[];
  language: string;
  songId: string;
  lyricsMarkdown: string;
  lyricsHtml: string;
  slug: string;
  __v: number;
}

export interface Book {
  _id: string;
  name: string;
  abbreviation: string;
  slug: string;
  __v: number;
}